+++ R/micromake.R
@@ -65,6 +65,12 @@ create_micro_file <- function(micro_var, x, filenames, keyword_values_long, out_
                           dplyr::anti_join(headings_with_only_na_values_long,
                                            by = c("variable-code", "code")
                                            ),
+           acrosscell = dplyr::select(x_new$acrosscell,
+                                      all_of(c(setdiff(names(new_data), figures_var),
+                                               names(get_base_acrosscell())
+                                               )
+                                             )
+                                      ),
            data       = new_data
            ) %>%
     fix_px() %>%