+++ NEWS.md
@@ -1,7 +1,7 @@
 # pxmake (development version)
 
 ## New features
-- Add modifying functions: `map()`, `baseperiod()`, `domain()`, `elimination()`, `descriptiondefault()`, `order()`, `precision()`.
+- Add modifying functions: `map()`, `baseperiod()`, `domain()`, `elimination()`, `descriptiondefault()`, `order()`, `precision()`, `cellnote()`, `cellnotex()`.
 - `micromake()` use furrr package to run in parallel. (#248)
 - `micromake()` removes headings if all figures are NA. (#250)